@@ -33,36 +33,33 @@
   },
   "homepage": "https://github.com/JoeKarlsson/bechdel-test",
   "scripts": {
+    "start": "node server/server.js",
     "test": "./node_modules/.bin/mocha -R spec tests/server/**/**.spec.js",
     "test:server:watch": "npm run test --watch",
     "test:client": "mocha tests/app/.setup.js tests/app/**/*-test.js",
-    "start": "node server/server.js",
+    "lint": "eslint src && npm run lint:css",
+    "lint:fix": "npm run lint -- --fix",
+    "lint:out": "npm run lint -- -f checkstyle -o coverage/checkstyle.xml",
+    "lint:css": "stylelint 'src/**/*.scss' --syntax scss",
+    "lint:css:fix": "stylelint 'src/**/*.scss' --syntax scss --fix",
     "build": "rimraf dist && set NODE_ENV=production && webpack --config ./server/webpack/webpack.production.config.js --progress --profile --colors",
     "eslint": "./node_modules/.bin/eslint .",
     "postinstall": "rimraf dist && set NODE_ENV=production && webpack --config ./server/webpack/webpack.production.config.js --progress --profile --colors"
   },
   "dependencies": {
-    "autoprefixer": "^6.0.3",
-    "babel-cli": "^6.4.0",
-    "babel-eslint": "^4.1.6",
-    "babel-jscs": "^2.0.5",
-    "babel-loader": "^6.2.3",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.1.0",
-    "babel-preset-stage-0": "^6.3.13",
+    "autoprefixer": "7.2.x",
     "bluebird": "^3.3.5",
     "body-parser": "^1.15.1",
-    "css-loader": "^0.19.0",
-    "d3": "^3.5.17",
-    "eslint": "^2.9.0",
-    "eslint-loader": "^1.3.0",
+    "css-loader": "0.28.x",
+    "eslint": "4.x.x",
+    "eslint-config-airbnb": "16.1.x",
+    "eslint-plugin-import": "2.8.x",
+    "eslint-plugin-jsx-a11y": "6.0.x",
+    "eslint-plugin-react": "7.4.x",
     "express": "^4.13.3",
     "express-favicon": "^1.0.1",
     "express-partials": "^0.3.0",
     "express-session": "^1.13.0",
-    "extract-text-webpack-plugin": "^0.8.2",
-    "html-webpack-plugin": "^1.6.1",
     "jquery": "^2.2.3",
     "json-loader": "^0.5.3",
     "method-override": "^2.3.5",
@@ -72,33 +69,30 @@
     "passport": "^0.3.2",
     "passport-github2": "^0.1.10",
     "postcss-loader": "^0.6.0",
-    "react": "^15.1.0",
-    "react-addons-css-transition-group": "^15.1.0",
-    "react-addons-test-utils": "^0.14.3",
-    "react-addons-transition-group": "^15.1.0",
-    "react-d3-library": "^1.1.4",
-    "react-dom": "^0.14.3",
-    "react-router": "^2.4.0",
-    "react-transform-hmr": "^1.0.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "request": "^2.67.0",
     "sass-loader": "^3.2.0",
     "sinon": "^1.17.4",
-    "stats-webpack-plugin": "^0.2.1",
     "style-loader": "^0.12.4",
-    "util": "^0.10.3",
-    "webpack": "^1.13.0"
+    "util": "^0.10.3"
   },
   "devDependencies": {
-    "babel-register": "^6.9.0",
+    "babel-core": "^6.26.0",
+    "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "6.24.x",
+    "babel-register": "6.x.x",
     "chai": "^3.5.0",
     "enzyme": "^2.3.0",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-config-standard": "^5.3.1",
-    "eslint-plugin-import": "^1.8.0",
-    "eslint-plugin-jsx-a11y": "^1.2.0",
-    "eslint-plugin-promise": "^1.1.0",
-    "eslint-plugin-react": "^5.0.1",
-    "eslint-plugin-standard": "^1.3.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "favicons-webpack-plugin": "0.0.7",
+    "html-webpack-plugin": "^1.7.0",
     "install": "^0.8.1",
     "jscs": "^2.1.1",
     "jsdom": "^6.5.1",
@@ -107,15 +101,23 @@
     "mocha-loader": "^0.7.1",
     "nodemon": "^1.9.2",
     "npm": "^3.9.5",
+    "prettier": "1.9.1",
+    "prop-types": "^15.6.0",
     "react-addons-test-utils": "^0.14.8",
     "rimraf": "^2.4.3",
     "sinon": "^1.16.1",
     "sinon-as-promised": "^4.0.0",
     "sinon-chai": "^2.8.0",
     "sinon-mongoose": "^1.2.1",
+    "stats-webpack-plugin": "^0.2.2",
+    "stylelint": "^8.3.1",
+    "stylelint-config-standard": "^18.0.0",
+    "stylelint-scss": "^2.2.0",
+    "stylelint-webpack-plugin": "^0.9.0",
     "supertest": "^1.2.0",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.2.0",
-    "webpack-node-externals": "^1.2.0"
+    "webpack": "^3.10.0",
+    "webpack-dev-middleware": "1.x.x",
+    "webpack-hot-middleware": "2.x.x",
+    "webpack-node-externals": "1.x.x"
   }
 }
