@@ -1,23 +1,24 @@
-import React from 'react';
-import { Link } from 'react-router';
-import FilmList from './FilmList.jsx';
-import skeleton from '../../shared/css/skeleton.css';
-import styles from './Films.scss';
-import * as $ from'jquery';
+import React from "react";
+import PropTypes from "prop-types";
+import { NavLink } from "react-router-dom";
+import FilmList from "./FilmList.jsx";
+import skeleton from "../../shared/css/skeleton.css";
+import styles from "./Films.scss";
+import * as $ from "jquery";
 
 class Films extends React.Component {
   constructor() {
     super();
     this.state = {
-      films: [],
+      films: []
     };
- }
+  }
 
   getAllFilms() {
     $.ajax({
-      url: '/api/film',
-      method: 'GET',
-      dataType: 'json',
+      url: "/api/film",
+      method: "GET",
+      dataType: "json",
       cache: false,
       success: function(data) {
         this.setState({ films: data });
@@ -36,36 +37,46 @@ class Films extends React.Component {
     return (
       <div className={styles.films}>
         <div className={skeleton.u_full_width}>
-          <div className='row'>
+          <div className="row">
             <h1>Bechdel.io</h1>
             <p>
-              The Bechdel.io was a collaborative project between a English Scholar and a Software Engineer. With the tool we’ve created, the process of determing whether a film passes the Beschdel Test is automated, which allows massive amounts of data to be generated with ease. Thus, data can be produced for large bodies of film, i.e. a certain director’s filmography, a certain actress’ body of work, or for the films released in a specific year.
+              The Bechdel.io was a collaborative project between a English
+              Scholar and a Software Engineer. With the tool we’ve created, the
+              process of determing whether a film passes the Beschdel Test is
+              automated, which allows massive amounts of data to be generated
+              with ease. Thus, data can be produced for large bodies of film,
+              i.e. a certain director’s filmography, a certain actress’ body of
+              work, or for the films released in a specific year.
             </p>
           </div>
         </div>
         <div className={styles.films_Header}>
           <h3>films</h3>
-          <Link to={'/film/new'}>
+          <NavLink to={"/film/new"}>
             <button>Add a Film</button>
-          </Link>
+          </NavLink>
         </div>
-        <div className='row'>
+        <div className="row">
           <FilmList films={this.state.films} />
         </div>
         <p>
-          If you have any suggestions for how to grow the application, we would love to hear from <a target="blank" href="https://www.callmejoe.net/contact/">you</a>.
+          If you have any suggestions for how to grow the application, we would
+          love to hear from{" "}
+          <a target="blank" href="https://www.callmejoe.net/contact/">
+            you
+          </a>.
         </p>
       </div>
-    )
+    );
   }
 }
 
 Films.propTypes = {
-  films: React.PropTypes.array,
+  films: PropTypes.array
 };
 
 Films.defaultProps = {
-  films: [],
+  films: []
 };
 
 export default Films;
