@@ -1,26 +1,20 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Router, Route, browserHistory, IndexRoute } from 'react-router';
-import normalize from './shared/css/normalize.css';
-import App from './App.jsx';
-import Home from './home/Home.jsx';
-import About from './static/about/About.jsx';
-import CaseStudy from './static/caseStudy/CaseStudy.jsx';
-import Films from './films/films/Films.jsx';
-import Film from './films/film/Film.jsx';
-import NewFilm from './films/new/NewFilm.jsx';
-import NoMatch from './shared/NoMatch.jsx';
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter as Router } from "react-router-dom";
+import App from "./App.jsx";
+import About from "./static/about/About.jsx";
+import CaseStudy from "./static/caseStudy/CaseStudy.jsx";
+import Films from "./films/films/Films.jsx";
+import Film from "./films/film/Film.jsx";
+import NewFilm from "./films/new/NewFilm.jsx";
+import NoMatch from "./shared/NoMatch.jsx";
+import ErrorBoundary from "./shared/ErrorBoundary/ErrorBoundary";
 
 ReactDOM.render(
-  <Router history={browserHistory}>
-    <Route path='/' component={App}>
-      <IndexRoute component={Films} />
-      <Route path='/about' component={About} />
-      <Route path='/case-study' component={CaseStudy} />
-      <Route path='/film/new' component={NewFilm} />
-      <Route path='/film/:id' component={Film} />
-      <Route path='*' component={NoMatch} />
-    </Route>
-  </Router>,
-  document.getElementById('root')
+  <ErrorBoundary>
+    <Router>
+      <App />
+    </Router>
+  </ErrorBoundary>,
+  document.getElementById("root")
 );
